[UITest] Use a difference logger name for each test
There are times when a test does not complete properly and the
logger is not disposed. Rest all the tests in the fixture start
failing because the logger has not been removed and disposed.
The fix is to make sure that each test creates a FileLogger with
the name being the same as the test name